make the port build on 10.9.5 .
provide a choice (subport) to install a build of the 4.1 API . It looks
like this should install in parallel to the 4.0 build but that remains
to be confirmed.

Currently it builds, links and installs, but currently with WEB_GL
disabled. That could be related to my custom Mesa port providing OpenEGL
and GLESv1,2 .
The MiniBrowser seems to work but hangs on opening about:gpu and Epiphany
seems to be unable to render pages at all (but without the UI lockup).
Both times I tried the new build I ended up getting a KP; once while
in the MiniBrowser, once after running Epiphany and (quite soon after)
a suspend/wake cycle.